refactor: light client (#1768)
* refactor: indexer in light-client

* refactor: light-client ProofRpcResultV2

* refactor: add from v1 proof result

* refactor: light-client unify get_validity_proof and v2

* refactor: light client unify get_compressed_account_by_owner and change return type to Account

* refactor: light-client get_compressed_token_accounts return type and unify v2

* refactor: get compressed token balances

* stopped compiling conversion.rs

* refactor: light-client add response types to Indexer trait

* stash

* stash

* fix: lint in light client

* fixed forester-utils

* program test compiles

* fix: light-test-utils

* fix: forester

* system-cpi-test compiles

* fix: system-test

* fixed: light-compressed-token

* fixed: program-tests

* fix: forester test and lint

* uncomment test-sbf feature

* remove unused conversions.rs

* fix: forester test and lint

* fix: sdk test

* fix: tests

* fix: get_validity_proof

* feat: extend Indexer trait

* stash

* simplify ItemsWithCursor

* fixed forester utils

* fixed: light-program-test, light-test-utils, forester-utils

* fix: tests

* chore: rename get_compressed_token_balances_by_owner -> get_compressed_token_balances_by_owner_v2

* use filter in get_compressed_accounts

* refactor: rename ValiditProofWithContext compressed_proof -> proof

* chore: add cpi context pubkeys to QUEUE_TREE_MAPPING and TreeInfo

* fix: compile error

+++ sdk-libs/photon-api/src/models/owner_balance.rs
@@ -13,14 +13,14 @@ use crate::models;
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct OwnerBalance {
     #[serde(rename = "balance")]
-    pub balance: u32,
+    pub balance: u64,
     /// A Solana public key represented as a base58 string.
     #[serde(rename = "owner")]
     pub owner: String,
 }
 
 impl OwnerBalance {
-    pub fn new(balance: u32, owner: String) -> OwnerBalance {
+    pub fn new(balance: u64, owner: String) -> OwnerBalance {
         OwnerBalance { balance, owner }
     }
 }
